refactor(dev): CTL-529 phase 5 — extract tailer.mjs, finalize thin index.mjs
Moves the reactive event-log follow layer (readNewEvents, startTailing,
loadExistingRegistrations + the four tailer state vars) into tailer.mjs,
and reduces index.mjs to its final shape: the daemon entrypoint
(main/shutdown, PID helpers, key-health) plus the complete re-export
barrel.

tailer.mjs exposes seedTailer({logPath,byteOffset}) so main() can seed the
follow offset without writing the module bindings directly, and
stopTailing() so shutdown() can close the fs.watch handle. main()'s
inline lastLogPath/lastByteOffset writes and shutdown()'s
eventsWatcher.close() are repointed through them — behavior-preserving.

shutdown() keeps the pre-existing BROKER_STATE_FILE dangling reference
byte-for-byte (plan "no behavior change" — Open Decisions Resolved #4);
correcting it would start removing broker.state.json on shutdown.

End state: index.mjs is 364 lines (was 2,390). The module DAG is acyclic
— config → state → projection → router → tailer → index — and index.mjs
is a pure sink (nothing imports it but the test files). All 55 public
symbols re-export from ./index.mjs.

Broker suite green: 273 pass / 0 fail across 7 files.
